requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
boto3==1.35.60
botocore==1.35.60
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cmake==3.25.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-inmemorystorage==2.1.0
Django==4.0.3
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-environ==0.9.0
django-resized==1.0.2
django-templated-mail==1.1.1
django-types==0.19.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djangorestframework-types==0.9.0
djoser==2.1.0
gunicorn==20.1.0
huey==2.4.4
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.0
Pillow==9.0.1
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2021.3
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.10.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
types-psycopg2==2.9.21.20241019
typing_extensions==4.12.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.11
